evidence: the network-filesystem check only compiles on the host that ships it
CI rejected an unnecessary int64 conversion in fs_linux.go. Statfs_t.Type is
already int64 there, so the conversion said nothing — and the reason it was
written at all is that this file never compiles on the machine it was written
on. A build-tagged file is only as reviewed as the platform that builds it.

Typing the constants instead would have moved the same problem onto an
architecture where that field is narrower. They are untyped now, so they adopt
whatever width the target has, and the switch reads the field directly.

Verified rather than assumed, since the local toolchain cannot see this file:
go vet for linux/amd64 and linux/arm64, and unconvert — the linter that
actually failed — built natively and pointed at the linux build, which is the
combination that works. golangci-lint cross-compiled from darwin panics inside
staticcheck's IR builder and cannot answer the question at all.

+++ evidence/sqlite/fs_linux.go
@@ -28,6 +28,10 @@ import (
 // exactly where SQLite's own documentation says not to put a database — POSIX
 // advisory locking over them is unreliable, and the failure is silent
 // corruption rather than an error.
+// Untyped, so they adopt whatever width Statfs_t.Type has on the target. It is
+// int64 on the platforms this project publishes for; converting it explicitly
+// is what CI rejected, and typing the constants would move the same problem
+// onto an architecture where the field is narrower.
 const (
 	nfsSuper  = 0x6969
 	smbSuper  = 0x517B
@@ -42,7 +46,7 @@ func refuseNetworkFilesystem(path string) error {
 		// costs an install; a missed one costs a database months later.
 		return nil //nolint:nilerr // see comment
 	}
-	switch int64(st.Type) {
+	switch st.Type {
 	case nfsSuper, smbSuper, cifsMagic, smb2Magic:
 		return fmt.Errorf(
 			"sqlite: %s is on a network filesystem, where SQLite's locking is unreliable and "+
